+++ activesupport/lib/active_support/core_ext/enumerable.rb
@@ -104,3 +104,17 @@ class Range #:nodoc:
     end
   end
 end
+
+# Array#sum was added in Ruby 2.4 but it only works with Numeric elements.
+#
+# We tried shimming it to attempt the fast native method, rescue TypeError,
+# and fall back to the compatible implementation, but that's much slower than
+# just calling the compat method in the first place.
+if Array.instance_methods(false).include?(:sum) && (%w[a].sum rescue true)
+  class Array
+    def sum(*args) #:nodoc:
+      # Use Enumerable#sum instead.
+      super
+    end
+  end
+end